+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.19
 
+* ``Signal`` improvements.
 * ``StateMachine`` now clears the ``onBeforeStateChange`` and ``onAfterStateChange`` upon calling ``clear()``
 
 ## 1.0.18